requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
livekit
livekit-agents
livekit-plugins-openai
livekit-plugins-deepgram
livekit-plugins-silero
python-multipart==0.0.9
